refactor: 로직 수정, 스피너 적용
